Task_1/requirements.txt
scapy>=2.5.0
dnspython>=2.6.1